Made the BEAKL layout default
Also simplified the configuration a bit:

- Put a simple shift key on the outer right thumb as I was getting a lot
of errors with the mod tap on pinkies.
- Backspace (left) and delete (right) now use the ring-index combo on
the default layer.
- All symbols are now on one layer, and workspace navigation is with
media keys. So left inner thumb is number, right inner thumb is symbols,
and both held together are FKeys. Outer left is navigation, but then
when you hold the outer right while in nav you get workspace and media.
- Parens are gone from the symbol layer to make room for all the symbols
on one layer, so these are on vertical combos between the top and middle
rows with opening symbols on left half and closing on right. Middle finger
is curly, index is round, and index on inner col is square.
- Got rid of numlock code.