Fixed reset key
Turns out that you need to properly set the bootloader in the rules.mk
file in order for the reset keycode to properly put the board into
bootloader mode. The default bootloader for the Sweep is caterina, but
the Elite-C needs atmel-dfu. Once this is set in the custom rules.mk,
then you can load up QMK Toolbox, hit the reset key (top left key on the
workspace navigation layer for me), and it will enter the bootloader so
you can press the 'Flash' button in the app.