Moved shift to index finger
So index is shift, middle is command, ring is alt, pinky is control. I'm
hoping that once I get used to it, this should make capitalisation a bit
more reliable as I'm using my strongest and fastest finger to shift.

Also moved some of the symbols on the num layer to line up with their
positions on the base layer to make it easier to remember. This may be a
zero sum game because the same symbols don't really line up with the
symbol layer, but I use the ones on the base layer more I guess.