fix: tweaked number and nav layers
Tried to make num layer more like a self-contained calculator layer,
with arithmetic operators, colon (for times), and shifted position of
zero and dot.

Also cleaned out the editing macros that are now on combos from nav,
and added a command+` window switcher on that layer. Leaving a cmd+tab
switcher off for now, as I am not sure how to implement that for ZMK
and I want them to match.